requirements.txt
defusedxml==0.7.1
Django==4.2.10
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.2
inflection==0.5.1
psycopg2-binary==2.9.9
six==1.16.0
sqlparse>=0.5.0 # not directly required, pinned by Snyk to avoid a vulnerability

